# docs/tech/api/openapi.yaml
openapi: 3.0.3
info:
  title: Pereobuyka API
  description: |
    HTTP API ядра «Переобуйка». Сценарии — docs/vision.md §4.
    LLM не подменяет факты; контекст подмешивается на сервере (vision §6).
  version: 1.0.0
servers:
  - url: /
...
paths:
...
  /api/v1/consultation/messages:
    get:
      tags: [Consultation]
      summary: История сообщений консультации
      operationId: listConsultationMessages
      security:
        - BearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsultationMessageListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [Consultation]
      summary: Отправить сообщение в LLM
      operationId: postConsultationMessage
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConsultationRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsultationResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /api/v1/consultation/transcribe:
    post:
      tags: [Consultation]
      summary: Распознать голосовой файл в текст (STT для Telegram)
      operationId: postConsultationTranscribe
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: Аудиофайл (например OGG/Opus из Telegram)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsultationTranscribeResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/ValidationError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    Unauthorized:
      description: Нет или невалидный токен
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    ValidationError:
      description: Ошибка валидации тела или параметров
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    ServiceUnavailable:
      description: Внешний сервис недоступен (например, LLM-провайдер)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties: true
...
    ConsultationRequest:
      type: object
      required: [message]
      properties:
        message:
          type: string
          maxLength: 4000
        history:
          type: array
          maxItems: 20
          items:
            $ref: "#/components/schemas/ConsultationHistoryItem"
          description: Предыдущие реплики user/assistant (для связного диалога в /ask)
    ConsultationHistoryItem:
      type: object
      required: [role, content]
      properties:
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
          maxLength: 4000
    ConsultationResponse:
      type: object
      required: [reply]
      properties:
        reply:
          type: string
        request_id:
          type: string
          format: uuid
          description: Идентификатор запроса для поддержки и диагностики
    ConsultationTranscribeResponse:
      type: object
      required: [text]
      properties:
        text:
          type: string
          description: Распознанный текст (для последующей отправки в POST /consultation/messages)
    ConsultationMessageOut:
      type: object
      required: [id, role, content, created_at]
      properties:
        id:
          type: string
          format: uuid
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
        created_at:
          type: string
          format: date-time
        request_id:
          type: string
          format: uuid
          nullable: true
    ConsultationMessageListResponse:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ConsultationMessageOut"
        total:
          type: integer
...